Organize entrypoints in cmd/ folder
