+++ cmd/operator/main.go
