+++ images/operator/Dockerfile
@@ -7,15 +7,16 @@ COPY go.mod go.sum /workspace/
 WORKDIR /workspace
 RUN go mod download
 
-COPY pkg/config /workspace/pkg/config
-COPY pkg/consts /workspace/pkg/consts
-COPY pkg/lib /workspace/pkg/lib
-COPY pkg/operator /workspace/pkg/operator
-COPY pkg/types /workspace/pkg/types
-COPY pkg/crds /workspace/pkg/crds
-COPY pkg/workloads /workspace/pkg/workloads
-WORKDIR /workspace/pkg/operator
-RUN GO111MODULE=on CGO_ENABLED=0 GOOS=linux go build -installsuffix cgo -o operator .
+COPY pkg/config pkg/config
+COPY pkg/consts pkg/consts
+COPY pkg/lib pkg/lib
+COPY pkg/operator pkg/operator
+COPY pkg/types pkg/types
+COPY pkg/crds pkg/crds
+COPY pkg/workloads pkg/workloads
+COPY cmd/operator cmd/operator
+
+RUN GO111MODULE=on CGO_ENABLED=0 GOOS=linux go build -installsuffix cgo -o operator ./cmd/operator
 
 
 FROM alpine:3.12
@@ -25,7 +26,7 @@ RUN chmod +x /usr/local/bin/kubectl
 
 RUN apk --no-cache add ca-certificates bash
 
-COPY --from=builder /workspace/pkg/operator/operator /root/
+COPY --from=builder /workspace/operator /root/
 RUN chmod +x /root/operator
 
 EXPOSE 8888
